+++ wayland-tests/tests/client_proxies.rs
@@ -10,7 +10,7 @@ use wayc::Proxy;
 #[test]
 fn proxy_equals() {
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ways::protocol::wl_compositor::WlCompositor>(1, ());
+    server.display.create_global::<ways::protocol::wl_compositor::WlCompositor>(1, ());
     let mut server_ddata = ServerHandler { output: None };
 
     let (_, mut client) = server.add_client();
@@ -55,7 +55,7 @@ fn proxy_equals() {
 #[test]
 fn proxy_user_data() {
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ways::protocol::wl_compositor::WlCompositor>(1, ());
+    server.display.create_global::<ways::protocol::wl_compositor::WlCompositor>(1, ());
     let mut server_ddata = ServerHandler { output: None };
 
     let (_, mut client) = server.add_client();
@@ -101,7 +101,7 @@ fn proxy_user_data() {
 #[test]
 fn dead_proxies() {
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ways::protocol::wl_output::WlOutput>(3, ());
+    server.display.create_global::<ways::protocol::wl_output::WlOutput>(3, ());
     let mut server_ddata = ServerHandler { output: None };
 
     let (_, mut client) = server.add_client();
@@ -145,8 +145,8 @@ fn dead_proxies() {
 #[test]
 fn dead_object_argument() {
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ways::protocol::wl_compositor::WlCompositor>(1, ());
-    server.display.handle().create_global::<ways::protocol::wl_output::WlOutput>(3, ());
+    server.display.create_global::<ways::protocol::wl_compositor::WlCompositor>(1, ());
+    server.display.create_global::<ways::protocol::wl_output::WlOutput>(3, ());
     let mut server_ddata = ServerHandler { output: None };
 
     let (_, mut client) = server.add_client();
@@ -196,7 +196,7 @@ impl ways::GlobalDispatch<ways::protocol::wl_output::WlOutput> for ServerHandler
     type GlobalData = ();
     fn bind(
         &mut self,
-        _: &mut ways::DisplayHandle<'_, Self>,
+        _: &mut ways::DisplayHandle<'_>,
         _: &ways::Client,
         output: ways::New<ways::protocol::wl_output::WlOutput>,
         _: &(),
@@ -215,7 +215,7 @@ impl ways::Dispatch<ways::protocol::wl_compositor::WlCompositor> for ServerHandl
         _: &ways::protocol::wl_compositor::WlCompositor,
         request: ways::protocol::wl_compositor::Request,
         _: &(),
-        dhandle: &mut ways::DisplayHandle<'_, Self>,
+        dhandle: &mut ways::DisplayHandle<'_>,
         data_init: &mut ways::DataInit<'_, Self>,
     ) {
         if let ways::protocol::wl_compositor::Request::CreateSurface { id } = request {