+++ wayland-tests/tests/globals.rs
@@ -12,7 +12,7 @@ use std::ops::Range;
 #[test]
 fn simple_global() {
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ServerCompositor>(1, ());
+    server.display.create_global::<ServerCompositor>(1, ());
 
     let (_, mut client) = server.add_client();
     let mut client_ddata = ClientHandler { globals: wayc::globals::GlobalList::new() };
@@ -34,10 +34,10 @@ fn simple_global() {
 #[test]
 fn multi_versions() {
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ServerCompositor>(4, ());
-    server.display.handle().create_global::<ServerCompositor>(3, ());
-    server.display.handle().create_global::<ServerCompositor>(2, ());
-    server.display.handle().create_global::<ServerCompositor>(1, ());
+    server.display.create_global::<ServerCompositor>(4, ());
+    server.display.create_global::<ServerCompositor>(3, ());
+    server.display.create_global::<ServerCompositor>(2, ());
+    server.display.create_global::<ServerCompositor>(1, ());
 
     let (_, mut client) = server.add_client();
     let mut client_ddata = ClientHandler { globals: wayc::globals::GlobalList::new() };
@@ -62,7 +62,7 @@ fn multi_versions() {
 #[test]
 fn dynamic_global() {
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ServerCompositor>(1, ());
+    server.display.create_global::<ServerCompositor>(1, ());
 
     let (_, mut client) = server.add_client();
     let mut client_ddata = ClientHandler { globals: wayc::globals::GlobalList::new() };
@@ -75,17 +75,17 @@ fn dynamic_global() {
     roundtrip(&mut client, &mut server, &mut client_ddata, &mut ServerHandler).unwrap();
     assert!(client_ddata.globals.list().len() == 1);
 
-    server.display.handle().create_global::<ServerShell>(1, ());
+    server.display.create_global::<ServerShell>(1, ());
 
     roundtrip(&mut client, &mut server, &mut client_ddata, &mut ServerHandler).unwrap();
     assert!(client_ddata.globals.list().len() == 2);
 
-    let output = server.display.handle().create_global::<ServerOutput>(1, ());
+    let output = server.display.create_global::<ServerOutput>(1, ());
 
     roundtrip(&mut client, &mut server, &mut client_ddata, &mut ServerHandler).unwrap();
     assert!(client_ddata.globals.list().len() == 3);
 
-    server.display.handle().remove_global(output);
+    server.display.remove_global(output);
 
     roundtrip(&mut client, &mut server, &mut client_ddata, &mut ServerHandler).unwrap();
     assert!(client_ddata.globals.list().len() == 2);
@@ -99,8 +99,8 @@ fn range_instantiate() {
         Proxy,
     };
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ServerCompositor>(4, ());
-    server.display.handle().create_global::<ServerShell>(1, ());
+    server.display.create_global::<ServerCompositor>(4, ());
+    server.display.create_global::<ServerShell>(1, ());
 
     let (_, mut client) = server.add_client();
     let mut client_ddata = ClientHandler { globals: wayc::globals::GlobalList::new() };
@@ -165,7 +165,7 @@ fn range_instantiate() {
 #[should_panic]
 fn wrong_version_create_global() {
     let server = TestServer::<ServerHandler>::new();
-    server.display.handle().create_global::<ServerCompositor>(42, ());
+    server.display.create_global::<ServerCompositor>(42, ());
 }
 
 #[test]
@@ -173,7 +173,7 @@ fn wrong_global() {
     use wayc::protocol::wl_output::WlOutput;
 
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ServerCompositor>(4, ());
+    server.display.create_global::<ServerCompositor>(4, ());
 
     let (_, mut client) = server.add_client();
     let mut client_ddata = ClientHandler { globals: wayc::globals::GlobalList::new() };
@@ -198,7 +198,7 @@ fn wrong_global_version() {
     use wayc::protocol::wl_compositor::WlCompositor;
 
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ServerCompositor>(1, ());
+    server.display.create_global::<ServerCompositor>(1, ());
 
     let (_, mut client) = server.add_client();
     let mut client_ddata = ClientHandler { globals: wayc::globals::GlobalList::new() };
@@ -222,7 +222,7 @@ fn invalid_global_version() {
     use wayc::protocol::wl_compositor::WlCompositor;
 
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ServerCompositor>(1, ());
+    server.display.create_global::<ServerCompositor>(1, ());
 
     let (_, mut client) = server.add_client();
     let mut client_ddata = ClientHandler { globals: wayc::globals::GlobalList::new() };
@@ -246,7 +246,7 @@ fn wrong_global_id() {
     use wayc::protocol::wl_compositor::WlCompositor;
 
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ServerCompositor>(1, ());
+    server.display.create_global::<ServerCompositor>(1, ());
 
     let (_, mut client) = server.add_client();
     let mut client_ddata = ClientHandler { globals: wayc::globals::GlobalList::new() };
@@ -270,7 +270,7 @@ fn two_step_binding() {
     use wayc::protocol::{wl_compositor::WlCompositor, wl_output::WlOutput};
 
     let mut server = TestServer::new();
-    server.display.handle().create_global::<ServerCompositor>(1, ());
+    server.display.create_global::<ServerCompositor>(1, ());
 
     let (_, mut client) = server.add_client();
     let mut client_ddata = ClientHandler { globals: wayc::globals::GlobalList::new() };
@@ -283,7 +283,7 @@ fn two_step_binding() {
     roundtrip(&mut client, &mut server, &mut client_ddata, &mut ServerHandler).unwrap();
 
     // add a new global while clients already exist
-    server.display.handle().create_global::<ServerOutput>(1, ());
+    server.display.create_global::<ServerOutput>(1, ());
 
     roundtrip(&mut client, &mut server, &mut client_ddata, &mut ServerHandler).unwrap();