+++ wayland-server/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - The `DelegateDispatch` mechanism is changed around an explicit trait-base extraction of module
   state from the main compositor state.
+- The `DisplayHandle` no longer has a type parameter
+- Global manipulation methods are moved from `DisplayHandle` to `Display`
 
 ## 0.30.0-alpha1